+++ Sources/libmcpp/libmain.c
@@ -762,7 +762,7 @@ static void mcpp_main( void)
                 }
                 if (keep_spaces && wrong_line && infile
                         && *(infile->bptr) != '\n' && *(infile->bptr) != EOS) {
-                    src_col = infile->bptr - infile->buffer;
+                    src_col = (int)(infile->bptr - infile->buffer);
                     /* Remember the current colums  */
                     break;                  /* Do sharp() now       */
                 }
